FILE STRUCTURE UPDATE: - File renames:   - server.js > app.js   - start.js > server.js   - utils.js > utils/throwErr.js
