+++ boilerplate-webapp/server/db/db.js
@@ -1,8 +1,7 @@
 const Sequelize = require('sequelize');
 
-const dbName = 'boilerplatewebapp';
 const db = new Sequelize(
-  process.env.DATABASE_URL || `postgres://localhost:5432/${dbName}`,
+  process.env.DATABASE_URL,
   {
     logging: false,
   }
