+++ boilerplate-webapp/server/utils/throwErr.js
@@ -1,11 +1,7 @@
-// reusable functions
-
 const throwErr = (status, message) => {
   const err = new Error(message);
   err.status = status;
   throw err;
 }
 
-module.exports = {
-  throwErr,
-};
+module.exports = throwErr;
